YH-1023: rename marketplace entity to resource
