+++ src/entities/resource/index.ts
