+++ src/entities/resource/model/types/resource.ts
