+++ src/entities/resource/ui/ResourceCard/ResourceCard.module.css
