+++ src/entities/resource/ui/ResourceCard/ResourceCard.tsx
@@ -5,7 +5,7 @@ import { Flex } from '@/shared/ui/Flex';
 import { ImageWithWrapper } from '@/shared/ui/ImageWithWrapper';
 import { Text } from '@/shared/ui/Text';
 
-import { Resource } from '@/entities/marketplace';
+import { Resource } from '@/entities/resource';
 
 import styles from './ResourceCard.module.css';
